refactor(nav): move angle_to_dest fn to coords mod
