+++ src/navigator_node/navigator_node/coords.py
@@ -6,6 +6,7 @@ Doing so lets us assign goals coordinates instead of trying to adjust using loca
 translations to the physical Rover.
 """
 
+import math
 import sys
 from math import sqrt
 
@@ -28,8 +29,8 @@ def coordinate_from_aruco_pose(
     we can use that and the distance to the marker to calculate the estimated
     coordinate.
     """
-    marker_position: Point = _pose.pose.position
-    marker_orientation = _pose.pose.orientation
+    _marker_position: Point = _pose.pose.position
+    _marker_orientation = _pose.pose.orientation
 
     llogger.error("coordinate estimation is unimplemented!")
     sys.exit(1)
@@ -40,10 +41,24 @@ def get_distance_to_marker(marker: PoseStamped) -> float:
     Given the pose information for an ArUco marker relative to the rover,
     calculate the distance to the marker
     """
-    marker_position: Point = marker.pose.point
+    marker_position: Point = marker.pose.position
     distance_x: float = marker_position.x  # forward/backward distance
     distance_y: float = marker_position.y  # left/right distance
 
     # now we need to calculate the distance to the marker by taking the hypotenuse
     distance_to_marker: float = sqrt(distance_x**2 + distance_y**2)
     return distance_to_marker
+
+
+def get_angle_to_dest(
+    dest_coord: GeoPoint, current_coord: GeoPointStamped
+) -> float:
+    """Calculate the angle from the robot to the destination."""
+    # TODO: Double check trigonometry
+    unnormalized_angle = math.atan2(
+        dest_coord.latitude - current_coord.position.latitude,
+        dest_coord.longitude - current_coord.position.longitude,
+    )
+    # Normalize the angle to be between -pi and pi
+    normalized_angle = (unnormalized_angle + math.pi) % (2 * math.pi) - math.pi
+    return normalized_angle
